requirements.txt
fastapi
uvicorn[standard]
openai
chromadb
langchain-text-splitters
pydantic
python-dotenv
httpx
requests
gradio>=6.0
